hackaton/requirements.txt
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
plotly-express==0.4.1
seaborn==0.13.2
streamlit==1.37.0
streamlit-option-menu==0.3.13
replicate==0.31.0
python-dotenv==1.0.0
matplotlib==3.9.1